Debugando

+++ Library/Models/Avaliacao.cs
@@ -13,13 +13,13 @@ namespace Library.Models
         private int      idProjeto;
         private DateTime dtAvaliacao;
         private string   feedback;
-        private double   notaAvaliacao;
+        private int   notaAvaliacao;
 
         public int IdAvaliacao { get => idAvaliacao; set => idAvaliacao = value; }
         public int IdEvento { get => idEvento; set => idEvento = value; }
         public int IdProjeto { get => idProjeto; set => idProjeto = value; }
         public DateTime DtAvaliacao { get => dtAvaliacao; set => dtAvaliacao = value; }
         public string Feedback { get => feedback; set => feedback = value; }
-        public double NotaAvaliacao { get => notaAvaliacao; set => notaAvaliacao = value; }
+        public int NotaAvaliacao { get => notaAvaliacao; set => notaAvaliacao = value; }
     }
 }
